add benchmarking script
